backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
pydantic-settings==2.8.1
passlib[bcrypt]==1.7.4
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
email-validator==2.2.0
opencv-python==4.11.0.86
numpy>=1.26,<2.2
pillow==11.1.0
ultralytics
